restructured cc
