--- include/cc/assert
@@ -1,3 +1,0 @@
-#pragma once
-
-#include <cc/detail/assert.hh>
