--- include/cc/breakpoint
@@ -1,3 +1,0 @@
-#pragma once
-
-#include <cc/detail/breakpoint.hh>
