--- include/cc/detail/error.hh
@@ -1,10 +1,0 @@
-#pragma once
-
-namespace cc::detail
-{
-template <class... E>
-struct error
-{
-    static constexpr bool value = false;
-};
-} // namespace cc::detail
