--- include/cc/detail/monostate.hh
@@ -1,8 +1,0 @@
-#pragma once
-
-namespace cc
-{
-struct monostate
-{
-};
-} // namespace cc
