--- include/cc/detail/traits/true_false_type.hh
@@ -1,13 +1,0 @@
-#pragma once
-
-namespace cc
-{
-struct true_type
-{
-    static inline constexpr bool value = true;
-};
-struct false_type
-{
-    static inline constexpr bool value = false;
-};
-} // namespace cc
