--- include/cc/fwd/less.hh
@@ -1,7 +1,0 @@
-#pragma once
-
-namespace cc
-{
-template <class T>
-struct less;
-}
