--- include/cc/fwd/optional.hh
@@ -1,9 +1,0 @@
-#pragma once
-
-namespace cc
-{
-struct nullopt_t;
-
-template <class T>
-struct optional;
-} // namespace cc
