--- include/cc/fwd/pair.hh
@@ -1,7 +1,0 @@
-#pragma once
-
-namespace cc
-{
-template <class A, class B>
-struct pair;
-}
