--- include/cc/fwd/unique_ptr.hh
@@ -1,7 +1,0 @@
-#pragma once
-
-namespace cc
-{
-template <class T>
-struct unique_ptr;
-}
