--- include/cc/fwd/vector.hh
@@ -1,7 +1,0 @@
-#pragma once
-
-namespace cc
-{
-template <class T>
-struct vector;
-}
