--- include/cc/fwds
@@ -1,7 +1,0 @@
-#pragma once
-
-#include <cc/fwd/optional.hh>
-#include <cc/fwd/pair.hh>
-#include <cc/fwd/span.hh>
-#include <cc/fwd/unique_ptr.hh>
-#include <cc/fwd/vector.hh>
