--- include/cc/hash
@@ -1,3 +1,0 @@
-#pragma once
-
-#include <cc/detail/hash.hh>
