--- include/cc/less
@@ -1,3 +1,0 @@
-#pragma once
-
-#include <cc/detail/less.hh>
