--- include/cc/macros
@@ -1,3 +1,0 @@
-#pragma once
-
-#include <cc/detail/macros.hh>
