--- include/cc/move
@@ -1,3 +1,0 @@
-#pragma once
-
-#include <cc/detail/move.hh>
