--- include/cc/optional
@@ -1,3 +1,0 @@
-#pragma once
-
-#include <cc/detail/optional.hh>
