--- include/cc/pair
@@ -1,3 +1,0 @@
-#pragma once
-
-#include <cc/detail/pair.hh>
