--- include/cc/span
@@ -1,3 +1,0 @@
-#pragma once
-
-#include <cc/detail/span.hh>
