--- include/cc/swap
@@ -1,3 +1,0 @@
-#pragma once
-
-#include <cc/detail/swap.hh>
