--- include/cc/typedefs
@@ -1,3 +1,0 @@
-#pragma once
-
-#include <cc/detail/typedefs.hh>
