--- include/cc/unique_ptr
@@ -1,3 +1,0 @@
-#pragma once
-
-#include <cc/detail/unique_ptr.hh>
