--- include/cc/vector
@@ -1,3 +1,0 @@
-#pragma once
-
-#include <cc/detail/vector.hh>
