+++ src/cc/always_false.hh
@@ -1,0 +1,10 @@
+#pragma once
+
+namespace cc
+{
+/// Helper class for indicating errors in static_asserts
+/// Usage:
+///   static_assert(cc::always_false<T>, "some error");
+template <class... E>
+constexpr bool always_false = false;
+}
