+++ src/cc/fwd.hh
@@ -1,0 +1,25 @@
+#pragma once
+
+namespace cc
+{
+template <class T>
+struct hash;
+template <class T>
+struct less;
+
+struct nullopt_t;
+template <class T>
+struct optional;
+
+template <class A, class B>
+struct pair;
+
+template <class T>
+struct span;
+
+template <class T>
+struct unique_ptr;
+
+template <class T>
+struct vector;
+}
