+++ src/cc/hash.hh
@@ -1,15 +1,15 @@
 #pragma once
 
-#include <cc/detail/error.hh>
-#include <cc/fwd/hash.hh>
-#include <cc/typedefs>
+#include <cc/always_false.hh>
+#include <cc/fwd.hh>
+#include <cc/typedefs.hh>
 
 namespace cc
 {
 template <class T>
 struct hash
 {
-    static_assert(detail::error<T>::value, "no hash<> specialization found");
+    static_assert(always_false<T>, "no hash<> specialization found");
 
     // hash_t operator()(T const& value) const noexcept { ... }
 };
@@ -20,4 +20,4 @@ struct hash<T*>
     [[nodiscard]] hash_t operator()(T* const& value) const noexcept { return hash_t(value); }
 };
 
-} // namespace cc
+}
