+++ src/cc/optional.hh
@@ -1,17 +1,17 @@
 #pragma once
 
-#include <cc/assert>
-#include <cc/fwd/hash.hh>
-#include <cc/fwd/less.hh>
-#include <cc/fwd/optional.hh>
-#include <cc/move>
+#include <cc/always_false.hh>
+#include <cc/assert.hh>
+#include <cc/fwd.hh>
+
+#include <utility>
 
 namespace cc
 {
 struct nullopt_t
 {
 };
-inline constexpr nullopt_t nullopt;
+constexpr nullopt_t nullopt;
 
 template <class T>
 struct optional
@@ -45,12 +45,12 @@ struct optional
     template <class U>
     [[nodiscard]] constexpr T const& value_or(U&& default_value) const&
     {
-        return _has_value ? _data.v : static_cast<T>(forward<U>(default_value));
+        return _has_value ? _data.v : static_cast<T>(std::forward<U>(default_value));
     }
     template <class U>
     [[nodiscard]] constexpr T&& value_or(U&& default_value) &&
     {
-        return _has_value ? static_cast<T&&>(_data.v) : static_cast<T>(forward<U>(default_value));
+        return _has_value ? static_cast<T&&>(_data.v) : static_cast<T>(std::forward<U>(default_value));
     }
 
     ~optional()
@@ -58,7 +58,6 @@ struct optional
         if (_has_value)
             _data.v.~T();
     }
-
 
 private:
     struct empty
@@ -74,11 +73,11 @@ private:
 template <class T>
 struct optional<T&>
 {
-    static_assert(sizeof(T) >= 0, "cannot form optional reference");
+    static_assert(always_false<T>, "cannot form optional reference");
 };
 template <class T>
 struct optional<T&&>
 {
-    static_assert(sizeof(T) >= 0, "cannot form optional reference");
+    static_assert(always_false<T>, "cannot form optional reference");
 };
-} // namespace cc
+}
